refactor(core): add utilities to create RichTextEditor views
